Comments: Also strip tags from body
Even with the changes from #39, we are still not producing a valid
feed since XML doesn't recognize our custom elements as valid.

This commit thus also strips the tags from the message body. It
would be nicer to transform them to valid HTML but that would also
require making sure no evil HTML is emitted in the feed and I don't
feel like implementing that right now. This works for now.